Changed price filed from int to double

+++ Module 2 - Object-Oriented Programming/1. The World of Objects/2. Constructors/Car.java
@@ -1,10 +1,10 @@
 public class Car {
     public String make;
-    public int price;
+    public double price;
     public int year;
     public String color;
 
-    public Car(String make, int price, int year, String color) {
+    public Car(String make, double price, int year, String color) {
         this.make = make;
         this.price = price;
         this.year = year;
